+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This section contains changes that have been committed but not yet released.
 ### Added
 
 - Add support for Event conferencing
+- Add support for Account deletion
 
 ### Changed